The Gamepad Drumkit (Version 2.0)
Here's my assignment with the related updates! This update takes feedback from the user testers, making the UI more directly relate to the controls and fixing several useability bugs (now, the left joystick does not audibly play samples when centered, and one no longer has to press every button on startup to initialize the instrument). Additionally, gyroscopic control was added to momentarily control a delay effect (x axis controls dry/wet, y axis delay time, and z axis feedback). I know in your previous feedback you mentioned the idea of being able to hold the controlled effect in it's place upon release, but after getting user feedback I feared that a user might accidentally enable this effect and not no how to return it to it's previous state. Thus, I decided to just make this effect momentary like the rest of the effects in this project.